backend/app/requirements.txt
alembic==1.15.1
fastapi==0.115.6
keras==3.4.1
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.4.2
# tensorflow
# tensorflow_intel

annotated-types==0.7.0
anyio==4.7.0
asyncio==3.4.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cramjam==2.9.0
docopt==0.6.2
exceptiongroup==1.2.2
fastparquet==2024.11.0
fsspec==2024.10.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
hdfs==2.7.3
httptools==0.6.4
idna==3.10
numpy==1.26.4
packaging==24.2
pandas==2.2.3
protobuf==5.29.1
py4j==0.10.9.7
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==10.0
